+++ docs/changes.rst
@@ -7,8 +7,15 @@ Changes in `atelier`
 Work in progress
 ================
 
-Trying to fix https://travis-ci.org/lino-framework/book/jobs/375579936
-The book does not seem to find the `objects.inv` for :ref:`atelier`.
+
+:func:`atelier.projects.get_project_info_from_mod` didn't yet work in
+environments without a local :xfile:`config.py` file.  Such projects
+have neither a :xfile:`tasks.py` file not a :xfile:`setup.py` file,
+but at least they have a :attr:`main_package
+<atelier.projects.Project.main_package>` (and that's what intersphinx
+needs).  This fixes :ticket:`2385` (intersphinx does not find the
+`objects.inv` for :ref:`atelier` on Travis).
+
 I added a logger info::
 
     logger.info("intersphinx_mapping set to {}".format(